splits the routes into two separate routes
